Fix tests & days
