+++ 2020/day3/day3_2.py
@@ -26,6 +26,6 @@ if __name__ == "__main__":
     PUZZLE_DIR = pathlib.Path(__file__).parent
 
     with open(PUZZLE_DIR / "test_input.txt", encoding="utf-8") as f:
-        instruction_list: list[str] = f.read().splitlines()
+        instruction_list: list[str] = f.readlines()
 
     print(f"Part 2: {part2(instruction_list)}")
