+++ 2020/day4/day4.py
@@ -20,7 +20,7 @@ def part1(passport_list):
 if __name__ == "__main__":
     PUZZLE_DIR = pathlib.Path(__file__).parent
 
-    with open(PUZZLE_DIR / "test_input.txt", encoding="utf-8") as f:
-        instruction_list: list[str] = f.read().splitlines()
+    with open(PUZZLE_DIR / "input.txt", encoding="utf-8") as f:
+        instruction_list: list[str] = f.readlines()
 
     print(f"Part 1: {part1(instruction_list)}")
