+++ 2020/day5/day5_2.py
@@ -33,8 +33,9 @@ def part2(seat_list):
 
     for i in range(13, 880):
         if i not in sorted(id_list):
-            print(i)
-    return sorted(id_list)
+            return i
+
+    raise Exception("Error")
 
 
 if __name__ == "__main__":
