--- 2022/day22/test_input_part2.txt
@@ -1,14 +1,0 @@
-        1111
-        1111
-        1111
-        1111
-222233334444
-222233334444
-222233334444
-222233334444
-        55556666
-        55556666
-        55556666
-        55556666
-
-10R5L5R10L4R5L5
