+++ 2022/day5/day5.py
@@ -1,8 +1,11 @@
 import pathlib
 
 
-def part1(instruction_list):
-    NB_STACKS = 9
+def part1(instruction_list, is_test):
+    if is_test:
+        NB_STACKS = 3
+    else:
+        NB_STACKS = 9
 
     stack: list[list[str]] = [[] for i in range(NB_STACKS)]
 
@@ -12,7 +15,7 @@ def part1(instruction_list):
             and not instruction.startswith(" 1   2   3")
             and instruction != ""
         ):
-            for i in range(0, NB_STACKS * 4, 4):
+            for i in range(0, len(instruction), 4):
                 if instruction[i : i + 3] == "   ":
                     continue
                 if instruction[i] == "[":
@@ -39,7 +42,7 @@ def part1(instruction_list):
 if __name__ == "__main__":
     PUZZLE_DIR = pathlib.Path(__file__).parent
 
-    with open(PUZZLE_DIR / "input.txt", encoding="utf-8") as f:
+    with open(PUZZLE_DIR / "test_input.txt", encoding="utf-8") as f:
         instruction_list: list[str] = f.read().splitlines()
 
-    print(f"Part 1: {part1(instruction_list)}")
+    print(f"Part 1: {part1(instruction_list, is_test=True)}")
